Smart_Attendance_System_Web/requirements.txt
flask==3.1.0
flask-login==0.6.3
mysql-connector-python==9.3.0
python-dotenv==1.1.0
